refactor: meta & desc for each page
