+++ app/(default)/test-case-generator/page.tsx
@@ -10,15 +10,15 @@ import { Testimonial } from "@/components/testimonial";
 import { testimonialData } from "@/components/utils/testimonial";
 import testAndStubsGen from "@/public/images/TestGenHighlighted.json";
 import { blocksData } from "@/components/utils/common";
-import { featuresData,faqQuestions } from "@/components/utils/stub-generation";
+import { featuresData,faqQuestions } from "@/components/utils/test-case-generator";
 import BannerBunny from "@/public/images/banner-bunny.png";
 import FAQ from "@/components/pillar-page/faq";
 const heroData = {
-  titleTop: "Simplifying Testing with ",
-  titleBottom: "Dynamic Test Data Generator",
-  subtitleTop: "Transforming User Scenarios into Reliable Test Data Instantly!",
+  titleTop: "Boost E2E Testing  ",
+  titleBottom: "with Keploy Test Case Generator! Learn How",
+  subtitleTop: "Avoid confusion and Duplication in Test Execution!",
   subtitleBottom:
-    "Turn User Interactions into Test Cases and Data Stubs that actually works! 🚀",
+    "Remove the duplicate and redundant test cases in minutes. 🚀",
   animationSrc: testAndStubsGen,
 };
 const ProblemBlocksData = {
@@ -28,9 +28,9 @@ const ProblemBlocksData = {
   blocksData: blocksData,
 };
 const FeaturesData = {
-  title: " With Keploy's Test Data Generation ",
+  title: "With Keploy's DeDuplication",
   subtitle:
-    "Transforming interactions into realistic, diverse, and customizable stubs for precise testing",
+    "Replace multiple tools in your test stack with one zero-code platform.Generic Features list/tables",
   highlightTitle: "eliminate",
   featuresData: featuresData,
 };
